+++ apps/examples/CHANGELOG.md
@@ -1,1 +1,8 @@
 # examples
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @devnepal/use-form@1.0.1