+++ packages/UseValidation/CHANGELOG.md
@@ -1,1 +1,8 @@
 # @devnepal/use-form
+
+## 1.0.1
+
+### Patch Changes
+
+- - Shown server errors for fields not registered
+  - prevented re-setting touched state when re-registering the validators